[REFACTOR]: DogCard 폴더구조 수정
