+++ components/DogCard/ContainerStyle.ts
