+++ components/DogCard/DogCardStyle.ts
